Remove comment about using idempotent_key
Supporting it would increase the complexity of this code, because the
idempotent_key is only useful if you're going to automatically retry
requests when they fail.

Given the low-volume of traffic, this increase in complexity is probably
not worthwhile.

We can revisit this decision if we see lots of network-related errors
when talking to Stripe.